requirements.txt
pandas
python-dotenv
openpyxl
secure-smtplib
cryptography